Replace "" with nil for Postgres UUIDs
When checking is-empty/not-empty UUIDs we compare against the empty string.
But an empty string is not a valid UUID so this comparison is useless for
Postgres where a UUID field cannot have such a value. Assuming that a UUID
field is only ever compared to the empty string in the context of
is-empty/non-empty checks, replacing it with nil is OK.

Fixes https://github.com/metabase/metabase/issues/13294